Changes made to file names and relocation
